ENFINNNNNNNN Test fonctionnel

+++ CommuMoney/CommuMoney.DAL/DAL/Depenses_DAL.cs
@@ -12,10 +12,10 @@ namespace CommuMoney.DAL.DAL
         public int ID { get; set; }
         public int ID_Personne { get; set; }
         public int ID_Projet { get; set; }
-        public float Montant { get; set; }
+        public double Montant { get; set; }
 
-        public Depenses_DAL(int id_personne, int id_projet, float montant) => (ID_Personne, ID_Projet, Montant) = (id_personne, id_projet, montant);
-        public Depenses_DAL(int id, int id_personne, int id_projet, float montant) => (ID, ID_Personne, ID_Projet, Montant) = (id, id_personne, id_projet, montant);
+        public Depenses_DAL(int id_personne, int id_projet, double montant) => (ID_Personne, ID_Projet, Montant) = (id_personne, id_projet, montant);
+        public Depenses_DAL(int id, int id_personne, int id_projet, double montant) => (ID, ID_Personne, ID_Projet, Montant) = (id, id_personne, id_projet, montant);
 
         #region Insert
         public void Insert(SqlConnection connexion)
