+++ CommuMoney/CommuMoney.DAL/Depot/DepensesDepot_DAL.cs
@@ -25,7 +25,7 @@ namespace CommuMoney.DAL.Depot
                 var depense = new Depenses_DAL(reader.GetInt32(0),
                                         reader.GetInt32(1),
                                         reader.GetInt32(2),
-                                        reader.GetFloat(3));
+                                        reader.GetDouble(3));
                 listeDepenses.Add(depense);
             }
 
@@ -49,7 +49,7 @@ namespace CommuMoney.DAL.Depot
             {
                 depense = new Depenses_DAL(reader.GetInt32(0),
                                         reader.GetInt32(1),
-                                        reader.GetFloat(2));
+                                        reader.GetDouble(2));
             }
             else
             {
